+++ table/CreateTable_test.go
@@ -1,4 +1,4 @@
-package structure
+package table
 
 import (
 	"log"
