+++ table/connect.go
@@ -1,4 +1,4 @@
-package structure
+package table
 
 import (
 	"database/sql"
