+++ table/insert.go
@@ -1,0 +1,6 @@
+package table
+
+
+func (table *Table) Insert(info interface{}) (*Table, error) {
+	return table.clone().Exec.Insert(info).Table, nil
+}
