+++ table/update_test.go
@@ -1,11 +1,11 @@
-package structure
+package table
 
 import (
 	"log"
 	"testing"
 )
 
-func TestTable_insert(t *testing.T) {
+func TestTable_update(t *testing.T) {
 	config := DBConfig{
 		UserName: "root",
 		UserPassword: "lotus20001006",
@@ -18,15 +18,7 @@ func TestTable_insert(t *testing.T) {
 		log.Println(err)
 		return
 	}
-	type hello2 struct {
-		Name string
-		Say string
-		Number int64
-	}
-	hhello := hello2{
-		"China",
-		"nihao",
-		3,
-	}
-	db.Table("hello2").Insert(&hhello)
+
+	db.Table("Hello").Where("number = ?", 3).Update("Name = ?", "OOOO")
+
 }
