+++ table/where_test.go
@@ -1,11 +1,12 @@
-package structure
+package table
 
 import (
 	"log"
 	"testing"
 )
 
-func TestTable_update(t *testing.T) {
+func TestTable_Where(t *testing.T) {
+
 	config := DBConfig{
 		UserName: "root",
 		UserPassword: "lotus20001006",
@@ -18,7 +19,5 @@ func TestTable_update(t *testing.T) {
 		log.Println(err)
 		return
 	}
-
-	db.Table("Hello").Where("number = ?", 3).Update("Name = ?", "OOOO")
-
+	db.Table("Hello").Where("number = ?", 3)
 }
